Fix URLs in a couple of "new upload" buttons
There are a couple of "new upload" buttons that aren't working, one in the menu bar and one that's in the header for the uploads page. The problem seems to be that they are pointing at `/uploads/new?community=icl` which doesn't work any more.

Fix by changing the URLs to `/uploads/new`.